+++ test/qunit-tests.js
@@ -2,7 +2,7 @@
  * NAME: qunit-test.js
  * DESC: tests for qunit-test.html (coded to my O365 Dev Site - YMMV)
  * AUTH: https://github.com/gitbrent/
- * DATE: Oct 03, 2017
+ * DATE: Oct 25, 2017
  *
  * HOWTO: Generate text tables for README etc.:
  * sprLib.list('Employees').getItems(['Id', 'Name', 'Badge_x0020_Number']).then(function(arrData){ console.log(getAsciiTableStr(arrData)) });
@@ -11,11 +11,14 @@
  // QUnit.test("QUnit Base Test", function(assert){ assert.ok( true === true, "Passed!" ); });
  */
 
+const BASEURL = _spPageContextInfo.siteServerRelativeUrl;
 const RESTROOT = '/sites/dev';
-const RESTDEMO = '/sites/demo';
+const RESTDEMO1 = '/sites/dev/sandbox';
+const RESTDEMO2 = '/sites/dev/sandbox/';
 //
 const ARR_NAMES_FIRST = ['Jack','Mark','CutiePie','Steve','Barry','Clark','Diana','Star','Luke','Captain'];
 const ARR_NAMES_LAST  = ['Septiceye','Iplier','Martzia','Rodgers','Allen','Kent','Prince','Lord','Skywalker','Marvel'];
+const LIST_GUID2 = '23846527-218a-43a2-b5c1-7b55b6feb1a3';
 //
 var gTestUserId = 9;
 
@@ -29,10 +32,12 @@ function getAsciiTableStr(arrayResults) {
 
 	let table = new AsciiTable().setHeading( arrColHeadings );
 
-	arrayResults.forEach(function(obj,idx){
+	arrayResults.forEach((obj,idx)=>{
 		let vals = [];
 		$.each(obj, function(key,val){
-			vals.push( ( typeof val === 'object' && key != '__metadata' ? JSON.stringify(val) : val ) );
+			if ( typeof val === 'object' && JSON.stringify(val).indexOf("__deferred") > 0 ) val = "[[__deferred]]";
+			//vals.push( ( typeof val === 'object' && key != '__metadata' ? JSON.stringify(val) : val ) );
+			vals.push( typeof val === 'object' ? JSON.stringify(val) : val );
 		});
 		table.addRow(vals);
 	});
@@ -62,6 +67,7 @@ QUnit.module( "Library OPTIONS" );
 QUnit.module( "LIST > COLS and INFO Methods" );
 // ================================================================================================
 {
+	// TEST: Using GUID
 	['Departments', 'Employees', 'Empty', '8fda2798-dbbc-497d-9840-df87b08e09c1']
 	.forEach(function(list,idx){
 		QUnit.test(`sprLib.list('${list}').cols()`, function(assert){
@@ -96,12 +102,11 @@ QUnit.module( "LIST > COLS and INFO Methods" );
 	});
 
 	// TEST: Using `baseUrl`
-	['Empty', 'Reports', '23846527-218a-43a2-b5c1-7b55b6feb1a3']
+	['Documents', 'Site Assets', 'Site Pages']
 	.forEach(function(list,idx){
-		QUnit.test(`sprLib.list('${list}').baseUrl('${RESTDEMO}').cols()`, function(assert){
+		QUnit.test(`sprLib.list({ name:'${list}', baseUrl:'${RESTDEMO1}' }).cols()`, function(assert){
 			var done = assert.async();
-			// TEST:
-			sprLib.list(list).baseUrl(RESTDEMO).cols()
+			sprLib.list({ name:list, baseUrl:RESTDEMO1 }).cols()
 			.then(function(arrColObjs){
 				assert.ok( arrColObjs.length > 0		,"Pass: arrColObjs.length = " + arrColObjs.length );
 				assert.ok( Object.keys(arrColObjs[0])	,"Pass: Object.keys().... = " + Object.keys(arrColObjs[0]).toString() );
@@ -115,10 +120,9 @@ QUnit.module( "LIST > COLS and INFO Methods" );
 			});
 		});
 
-		QUnit.test(`sprLib.list('${list}').baseUrl('${RESTDEMO}').info()`, function(assert){
+		QUnit.test(`sprLib.list({ name:'${list}', baseUrl:'${RESTDEMO2}' }).info()`, function(assert){
 			var done = assert.async();
-			// TEST:
-			sprLib.list(list).baseUrl(RESTDEMO).info()
+			sprLib.list({ name:list, baseUrl:RESTDEMO2 }).info()
 			.then(function(objInfo){
 				assert.ok( objInfo.Id		     , "Pass: Id....... = " + objInfo.Id );
 				assert.ok( objInfo.Created	     , "Pass: Created.. = " + objInfo.Created );
@@ -178,20 +182,21 @@ QUnit.module( "LIST > COLS and INFO Methods" );
 QUnit.module( "LIST > BASEURL Methods" );
 // ================================================================================================
 {
-	QUnit.test(`sprLib.list('Empty').baseUrl('${RESTROOT}')`, function(assert){
-		var sprResult = sprLib.list('Employees').baseUrl(RESTROOT);
-		//
-		assert.ok( sprResult							, "Pass: sprResult.............................. = "+ sprResult );
-		assert.ok( typeof sprResult === 'object'		, "Pass: typeof sprResult === 'object'.......... = "+ typeof sprResult );
-		assert.ok( typeof sprResult.cols === 'function'	, "Pass: typeof sprResult.cols === 'function'... = "+ typeof sprResult.cols );
-		assert.ok( sprResult.baseUrl() == RESTROOT		, "Pass: sprResult.baseUrl() == RESTROOT........ = "+ sprResult.baseUrl() );
-	});
+	QUnit.test(`sprLib.list() 'baseUrl' parsing VS!`, function(assert){
+		var done = assert.async();
+		Promise.all([
+			sprLib.list({ name:'Employees', baseUrl:'/sites/dev//'}).cols(),
+			sprLib.list({ name:'Employees', baseUrl:'/sites/dev/' }).cols(),
+			sprLib.list({ name:'Employees', baseUrl:'/sites/dev'  }).cols()
+		])
+		.then(arrArr => {
+			assert.ok(arrArr[0].length > 0);
+			assert.ok(arrArr[1].length > 0);
+			assert.ok(arrArr[2].length > 0);
+			assert.ok( arrArr[0].length == arrArr[1].length && arrArr[1].length == arrArr[2].length, "Pass: lengths all the same: "+ arrArr[0].length+'/'+arrArr[1].length+'/'+arrArr[2].length );
 
-	QUnit.test(`sprLib.list('Empty').baseUrl() VS!`, function(assert){
-		var sprHas = sprLib.list('Employees').baseUrl('/sites/demo/').baseUrl();
-		var sprNot = sprLib.list('Employees').baseUrl('/sites/demo').baseUrl();
-		//
-		assert.ok( sprHas == sprNot, "Pass: sprHas == sprNot : '"+ sprHas +"' == '"+ sprNot+ "'" );
+			done();
+		})
 	});
 }
 
@@ -243,7 +248,7 @@ QUnit.module( "LIST > ITEM CRUD Methods" );
 		var done = assert.async();
 
 		// PREP:
-		sprLib.list('Employees').getItems({ listCols:'Id', queryOrderby:'Modified', queryLimit:1 })
+		sprLib.list('Employees').getItems({ listCols:'Id', queryOrderby:'Modified', queryLimit:1, metadata:true })
 		.then(function(data){
 			// TEST:
 			sprLib.list('Employees')
@@ -312,7 +317,7 @@ QUnit.module( "LIST > ITEM CRUD Methods" );
 		var done = assert.async();
 		// PREP:
 		var gUpdateItem = {};
-		sprLib.list('Employees').getItems({ listCols:'Id', queryOrderby:'Modified', queryLimit:1 })
+		sprLib.list('Employees').getItems({ listCols:'Id', queryOrderby:'Modified', queryLimit:1, metadata:true })
 		.then(function(data){ gUpdateItem = data[0]; })
 		.then(function(){
 			// TEST:
@@ -519,7 +524,7 @@ QUnit.module( "LIST > ITEM GET Methods" );
 		.getItems()
 		.then(function(arrayResults){
 			assert.ok( arrayResults.length > 0        , "arrayResults is an Array and length > 0: "+ arrayResults.length );
-			assert.ok( (arrayResults[0].__metadata.id), "arrayResults[0].__metadata.id exists: "+ JSON.stringify(arrayResults[0].__metadata.id) );
+			assert.ok( (!arrayResults[0].__metadata)  , "arrayResults[0].__metadata does not exist!" );
 			assert.ok( getAsciiTableStr(arrayResults) , `RESULTS:\n${getAsciiTableStr(arrayResults)}`);
 			done();
 		})
@@ -536,8 +541,7 @@ QUnit.module( "LIST > ITEM GET Methods" );
 		.getItems('Name')
 		.then(function(arrayResults){
 			assert.ok( arrayResults.length > 0                 , "arrayResults is an Array and length > 0: "+ arrayResults.length );
-			assert.ok( (arrayResults[0].__metadata.id)         , "arrayResults[0].__metadata.id exists: "+ JSON.stringify(arrayResults[0].__metadata.id) );
-			assert.ok( Object.keys(arrayResults[0]).length == 2, "arrayResults[0] has length == 2: "+ Object.keys(arrayResults[0]).length );
+			assert.ok( Object.keys(arrayResults[0]).length == 1, "arrayResults[0] has length == 1: "+ Object.keys(arrayResults[0]).length );
 			assert.ok( getAsciiTableStr(arrayResults)          , `RESULTS:\n${getAsciiTableStr(arrayResults)}`);
 			done();
 		})
@@ -556,8 +560,7 @@ QUnit.module( "LIST > ITEM GET Methods" );
 		)
 		.then(function(arrayResults){
 			assert.ok( arrayResults.length > 0                 , "arrayResults is an Array and length > 0: "+ arrayResults.length );
-			assert.ok( (arrayResults[0].__metadata.id)         , "arrayResults[0].__metadata.id exists: "+ JSON.stringify(arrayResults[0].__metadata.id) );
-			assert.ok( Object.keys(arrayResults[0]).length == 4, "arrayResults[0] has length == 4: "+ Object.keys(arrayResults[0]).length );
+			assert.ok( Object.keys(arrayResults[0]).length == 3, "arrayResults[0] has length == 3: "+ Object.keys(arrayResults[0]).length );
 			assert.ok( getAsciiTableStr(arrayResults)          , `RESULTS:\n${getAsciiTableStr(arrayResults)}`);
 			done();
 		})
@@ -573,8 +576,7 @@ QUnit.module( "LIST > ITEM GET Methods" );
 		sprLib.list('Employees')
 		.getItems({ listCols:'Manager/Title' })
 		.then(function(arrayResults){
-			assert.ok( Object.keys(arrayResults[0]).length == 2, "arrayResults[0] has length == 2: "+ Object.keys(arrayResults[0]).length );
-			assert.ok( (arrayResults[0].__metadata.id)         , "arrayResults[0].__metadata.id exists: "+ JSON.stringify(arrayResults[0].__metadata.id) );
+			assert.ok( Object.keys(arrayResults[0]).length == 1, "arrayResults[0] has length == 1: "+ Object.keys(arrayResults[0]).length );
 			assert.ok( (arrayResults[0].Manager.Title)         , "arrayResults[0].Manager.Title exists: "+ JSON.stringify(arrayResults[0].Manager.Title) );
 			assert.ok( getAsciiTableStr(arrayResults)          , `RESULTS:\n${getAsciiTableStr(arrayResults)}`);
 			done();
@@ -593,8 +595,7 @@ QUnit.module( "LIST > ITEM GET Methods" );
 			listCols:['Id', 'Name', 'Manager/Title']
 		})
 		.then(function(arrayResults){
-			assert.ok( Object.keys(arrayResults[0]).length == 4, "arrayResults[0] has length == 4: "+ Object.keys(arrayResults[0]).length );
-			assert.ok( (arrayResults[0].__metadata.id)         , "arrayResults[0].__metadata.id exists: "+ JSON.stringify(arrayResults[0].__metadata.id) );
+			assert.ok( Object.keys(arrayResults[0]).length == 3, "arrayResults[0] has length == 3: "+ Object.keys(arrayResults[0]).length );
 			assert.ok( (arrayResults[0].Manager.Title)         , "arrayResults[0].Manager.Title exists: "+ JSON.stringify(arrayResults[0].Manager.Title) );
 			assert.ok( getAsciiTableStr(arrayResults)          , `RESULTS:\n${getAsciiTableStr(arrayResults)}`);
 			done();
@@ -608,9 +609,10 @@ QUnit.module( "LIST > ITEM GET Methods" );
 	QUnit.test("sprLib.getItems() 06: `listCols` with simple array of col names (Manager/Id and Manager/Title)", function(assert){
 		var done = assert.async();
 		// TEST:
-		sprLib.list('Employees')
+		sprLib.list({ name:'Employees' })
 		.getItems({
-			listCols:['Id', 'Name', 'Manager/Id', 'Manager/Title']
+			listCols: ['Id', 'Name', 'Manager/Id', 'Manager/Title'],
+			metadata: true
 		})
 		.then(function(arrayResults){
 			assert.ok( Object.keys(arrayResults[0]).length == 4, "arrayResults[0] has length == 4: "+ Object.keys(arrayResults[0]).length );
@@ -636,7 +638,8 @@ QUnit.module( "LIST > ITEM GET Methods" );
 				badgeNum: { dataName:'Badge_x0020_Number' },
 				mgrId:    { dataName:'Manager/Id'         },
 				mgrTitle: { dataName:'Manager/Title'      }
-			}
+			},
+			metadata: true
 		})
 		.then(function(arrayResults){
 			assert.ok( arrayResults.length > 0                     , "arrayResults is an Array and length > 0: "+ arrayResults.length );
@@ -666,7 +669,7 @@ QUnit.module( "LIST > ITEM GET Methods" );
 			}
 		})
 		.then(function(arrayResults){
-			assert.ok( Object.keys(arrayResults[0]).length == 5, "arrayResults[0] has length == 5: "+ Object.keys(arrayResults[0]).length );
+			assert.ok( Object.keys(arrayResults[0]).length == 4, "arrayResults[0] has length == 4: "+ Object.keys(arrayResults[0]).length );
 			assert.ok( typeof arrayResults[0].mgrTitle === 'string', "`mgrTitle` named column returned a string value: "+ arrayResults[0].mgrTitle );
 			assert.ok(
 				'"'+arrayResults[0].funcTest+'"' == '"'+arrayResults[0].name+':'+arrayResults[0].badgeNum+'"',
@@ -695,7 +698,7 @@ QUnit.module( "LIST > ITEM GET Methods" );
 			queryFilter: "Departments_x0020_Supported ne null"
 		})
 		.then(function(arrayResults){
-			assert.ok( Object.keys(arrayResults[0]).length == 4, "arrayResults[0] has length == 4: "+ Object.keys(arrayResults[0]).length );
+			assert.ok( Object.keys(arrayResults[0]).length == 3, "arrayResults[0] has length == 3: "+ Object.keys(arrayResults[0]).length );
 			assert.ok( !isNaN(arrayResults[0].depsArr[0].Id), "arrayResults[0].depsArr[0].Id is a number: "+ arrayResults[0].depsArr[0].Id );
 			assert.ok( getAsciiTableStr(arrayResults), `RESULTS:\n${getAsciiTableStr(arrayResults)}`);
 			done();
@@ -716,7 +719,7 @@ QUnit.module( "LIST > ITEM GET Methods" );
 			queryFilter: "Departments_x0020_Supported ne null"
 		})
 		.then(function(arrayResults){
-			assert.ok( Object.keys(arrayResults[0]).length == 4                  , "arrayResults[0] has length == 4: "+ Object.keys(arrayResults[0]).length );
+			assert.ok( Object.keys(arrayResults[0]).length == 3                  , "arrayResults[0] has length == 3: "+ Object.keys(arrayResults[0]).length );
 			assert.ok( (arrayResults[0].Manager.Title)                           , "arrayResults[0].Manager.Title exists: "+ JSON.stringify(arrayResults[0].Manager.Title) );
 			assert.ok( Array.isArray(arrayResults[0].Departments_x0020_Supported), "arrayResults[0].Dep[..]ted is an array: true" );
 			assert.ok( (arrayResults[0].Departments_x0020_Supported[0].Id)       , "arrayResults[0].Dep[..]ted[0].Id exists: "+ JSON.stringify(arrayResults[0].Departments_x0020_Supported[0].Id) );
@@ -750,7 +753,7 @@ QUnit.module( "LIST > ITEM GET Methods" );
 		])
 		.then(function(arrayResults){
 			var res1 = arrayResults[0];
-			assert.ok( Object.keys(res1[0]).length == 3, "res1[0] has length == 3: "+ Object.keys(res1[0]).length );
+			assert.ok( Object.keys(res1[0]).length == 2, "res1[0] has length == 2: "+ Object.keys(res1[0]).length );
 			assert.ok(
 				!isNaN( res1[0].Departments_x0020_Supported[0].Id ),
 				"res1[0].Departments_x0020_Supported[0].Id is a number: "+ res1[0].Departments_x0020_Supported[0].Id
@@ -763,7 +766,7 @@ QUnit.module( "LIST > ITEM GET Methods" );
 
 			// Empty Multi-Lookup fields should be an empty array `[]`
 			var res2 = arrayResults[1];
-			assert.ok( Object.keys(res2[0]).length == 3, "res2[0] has length == 3: "+ Object.keys(res2[0]).length );
+			assert.ok( Object.keys(res2[0]).length == 2, "res2[0] has length == 2: "+ Object.keys(res2[0]).length );
 			assert.ok(
 				(Array.isArray(res2[0].Departments_x0020_Supported) && res2[0].Departments_x0020_Supported.length == 0),
 				"res2[0].Departments_x0020_Supported == []: "+ res2[0].Departments_x0020_Supported
@@ -796,14 +799,14 @@ QUnit.module( "LIST > ITEM GET Methods" );
 		])
 		.then(function(arrayResults){
 			var res1 = arrayResults[0];
-			assert.ok( Object.keys(res1[0]).length == 3, "res1[0] has length == 3: "+ Object.keys(res1[0]).length );
+			assert.ok( Object.keys(res1[0]).length == 2, "res1[0] has length == 2: "+ Object.keys(res1[0]).length );
 			assert.ok( !isNaN(res1[0].Mentored_x0020_Team_x0020_Member[0].ID), "res1[0].Mentored_x0020_Team_x0020_Member[0].ID is a number: "+ res1[0].Mentored_x0020_Team_x0020_Member[0].ID );
 			assert.ok( (res1[0].Mentored_x0020_Team_x0020_Member[0].Title), "res1[0].Mentored_x0020_Team_x0020_Member[0].Title exists: "+ res1[0].Mentored_x0020_Team_x0020_Member[0].Title );
 			assert.ok( getAsciiTableStr(res1), `RESULTS:\n${getAsciiTableStr(res1)}`);
 
 			// Empty Person/Lookup fields should be null
 			var res2 = arrayResults[1];
-			assert.ok( Object.keys(res2[0]).length == 3, "res2[0] has length == 3: "+ Object.keys(res2[0]).length );
+			assert.ok( Object.keys(res2[0]).length == 2, "res2[0] has length == 2: "+ Object.keys(res2[0]).length );
 			assert.ok( (res2[0].Mentored_x0020_Team_x0020_Member == null), "res2[0].Mentored_x0020_Team_x0020_Member == null: "+ res2[0].Mentored_x0020_Team_x0020_Member );
 			assert.ok( getAsciiTableStr(res2), `RESULTS:\n${getAsciiTableStr(res2)}`);
 
@@ -830,7 +833,7 @@ QUnit.module( "LIST > ITEM GET Methods" );
 				queryFilter: 'Id eq '+intId
 			})
 			.then(function(arrayResults){
-				assert.ok( Object.keys(arrayResults[0]).length == 3, "arrayResults[0] has length == 3: "+ Object.keys(arrayResults[0]).length );
+				assert.ok( Object.keys(arrayResults[0]).length == 2, "arrayResults[0] has length == 2: "+ Object.keys(arrayResults[0]).length );
 				assert.ok( (arrayResults[0].Id == intId)           , "arrayResults[0].Id == intId: "+ intId );
 				assert.ok( (arrayResults[0].Name)                  , "arrayResults[0].Name exists: "+ JSON.stringify(arrayResults[0].Name) );
 				assert.ok( getAsciiTableStr(arrayResults)          , `RESULTS:\n${getAsciiTableStr(arrayResults)}`);
@@ -859,7 +862,7 @@ QUnit.module( "LIST > ITEM GET Methods" );
 				queryLimit: 1
 			})
 			.then(function(arrayResults){
-				assert.ok( Object.keys(arrayResults[0]).length == 2, "arrayResults[0] has length == 2: "+ Object.keys(arrayResults[0]).length );
+				assert.ok( Object.keys(arrayResults[0]).length == 1, "arrayResults[0] has length == 1: "+ Object.keys(arrayResults[0]).length );
 				assert.ok( (arrayResults[0].Name)                  , "arrayResults[0].Name exists: "+ JSON.stringify(arrayResults[0].Name) );
 				assert.ok( getAsciiTableStr(arrayResults)          , `RESULTS:\n${getAsciiTableStr(arrayResults)}`);
 				done();
@@ -887,7 +890,7 @@ QUnit.module( "LIST > ITEM GET Methods" );
 				queryLimit: 1
 			})
 			.then(function(arrayResults){
-				assert.ok( Object.keys(arrayResults[0]).length == 2, "arrayResults[0] has length == 2: "+ Object.keys(arrayResults[0]).length );
+				assert.ok( Object.keys(arrayResults[0]).length == 1, "arrayResults[0] has length == 1: "+ Object.keys(arrayResults[0]).length );
 				assert.ok( (arrayResults[0].Name)                  , "arrayResults[0].Name exists: "+ JSON.stringify(arrayResults[0].Name) );
 				assert.ok( getAsciiTableStr(arrayResults)          , `RESULTS:\n${getAsciiTableStr(arrayResults)}`);
 				done();
@@ -903,7 +906,7 @@ QUnit.module( "LIST > ITEM GET Methods" );
 		var done = assert.async();
 		// TEST:
 		sprLib.list('Employees')
-		.getItems({ listCols:'Name', queryLimit:3 })
+		.getItems({ listCols:'Name', queryLimit:3, metadata:true })
 		.then(function(arrayResults){
 			assert.ok( arrayResults.length == 3       , "arrayResults length == 3: "+ arrayResults.length );
 			assert.ok( (arrayResults[0].__metadata.id), "arrayResults[0].__metadata.id exists: "+ JSON.stringify(arrayResults[0].__metadata.id) );
@@ -1008,19 +1011,19 @@ QUnit.module( "LIST > ITEM GET Methods" );
 			.then(function(arrayResults){
 				var result = arrayResults[0];
 				assert.ok( true, "Negative Test: getVersions=false\n------------------" );
-				assert.ok( Object.keys(result[0]).length == 4, "result[0] has length == 4: "+ Object.keys(result[0]).length );
+				assert.ok( Object.keys(result[0]).length == 3, "result[0] has length == 43: "+ Object.keys(result[0]).length );
 				assert.ok( (true), "result[0].Versioned_x0020_Comments: "+ result[0].Versioned_x0020_Comments );
 				assert.ok( getAsciiTableStr(result), `RESULTS:\n${getAsciiTableStr(result)}` );
 
 				var result = arrayResults[1];
 				assert.ok( true, "TEST: getVersions with keyname-1\n------------------" );
-				assert.ok( Object.keys(result[0]).length == 2, "result[0] has length == 2: "+ Object.keys(result[0]).length );
+				assert.ok( Object.keys(result[0]).length == 1, "result[0] has length == 1: "+ Object.keys(result[0]).length );
 				assert.ok( (true), "result[0].appendText: "+ result[0].appendText );
 				assert.ok( getAsciiTableStr(result), `RESULTS:\n${getAsciiTableStr(result)}` );
 
 				var result = arrayResults[2];
 				assert.ok( true, "TEST: getVersions with keyname-2\n------------------" );
-				assert.ok( Object.keys(result[0]).length == 2, "result[0] has length == 2: "+ Object.keys(result[0]).length );
+				assert.ok( Object.keys(result[0]).length == 1, "result[0] has length == 1: "+ Object.keys(result[0]).length );
 				assert.ok( (true), "result[0].Versioned_x0020_Comments: "+ result[0].Versioned_x0020_Comments );
 				assert.ok( (true), "result[0].VC is Array(): "+ Array.isArray(result[0].Versioned_x0020_Comments) );
 				assert.ok( getAsciiTableStr(result), `RESULTS:\n${getAsciiTableStr(result)}` );
@@ -1043,7 +1046,6 @@ QUnit.module( "LIST > ITEM GET Methods" );
 			sprLib.list('Employees').getItems(data)
 			.then(function(arrayResults){
 				assert.ok( arrayResults.length > 0        , "arrayResults is an Array and length > 0: "+ arrayResults.length );
-				assert.ok( (arrayResults[0].__metadata.id), "arrayResults[0].__metadata.id exists: "+ JSON.stringify(arrayResults[0].__metadata.id) );
 				assert.ok( getAsciiTableStr(arrayResults) , `RESULTS:\n${getAsciiTableStr(arrayResults)}`);
 				done();
 			})
@@ -1066,6 +1068,62 @@ QUnit.module( "LIST > ITEM GET Methods" );
 		console.log( arrayResults );
 	});
 	*/
+}
+
+// ================================================================================================
+QUnit.module( "QA: Result Parsing" );
+// ================================================================================================
+{
+	QUnit.test("sprLib.rest() ex: 'Parsing Lookups: Lookup with 2 sub items (ex: 'Member.ID' and 'Member.Title') - plain cols", function(assert){
+		var done = assert.async();
+		// TEST:
+		sprLib.rest({
+			url: '_api/web/roleAssignments',
+			queryCols: ['PrincipalId','Member/PrincipalType','Member/Title','RoleDefinitionBindings/Name','RoleDefinitionBindings/Hidden']
+		})
+		.then(function(arrayResults){
+			var objRow = arrayResults[0];
+			assert.ok( Object.keys(objRow).length == 3, "objRow has length == 3: "+ Object.keys(objRow).length );
+			assert.ok( (objRow.Member.PrincipalType && objRow.Member.Title), "objRow.Member.PrincipalType/Title both exist: "+ objRow.Member.PrincipalType+' / '+objRow.Member.Title );
+			assert.ok( (typeof objRow.Member === 'object' && !Array.isArray(objRow.Member)), "objRow.Member is object (and !array): "+ typeof objRow.Member === 'object' );
+			assert.ok( (Array.isArray(objRow.RoleDefinitionBindings)), "objRow.RoleDef is an array: "+ Array.isArray(objRow.RoleDefinitionBindings) );
+			assert.ok( (objRow.RoleDefinitionBindings[0].Name && objRow.RoleDefinitionBindings[0].Hidden !== 'undefined'), "objRow.RoleDef[0].Name/Hidden both exist: "+ objRow.RoleDefinitionBindings[0].Name+' / '+objRow.RoleDefinitionBindings[0].Hidden );
+			assert.ok( getAsciiTableStr(arrayResults), `RESULTS:\n${getAsciiTableStr(arrayResults)}` );
+			done();
+		})
+		.catch(function(err){
+			assert.ok( (false), err );
+			done();
+		});
+	});
+
+	QUnit.test("sprLib.rest() ex: 'Parsing Lookups: Lookup with 2 sub items (ex: 'Member.ID' and 'Member.Title') - col objects", function(assert){
+		var done = assert.async();
+		// TEST:
+		sprLib.rest({
+			url: '_api/web/roleAssignments',
+			queryCols: {
+				PrincipalId:	{ dataName:'PrincipalId' },
+				PrincipalType:	{ dataName:'Member/PrincipalType' },
+				Title:			{ dataName:'Member/Title' },
+				RoleNames:		{ dataName:'RoleDefinitionBindings/Name' }
+			}
+		})
+		.then(function(arrayResults){
+			var objRow = arrayResults[0];
+			assert.ok( Object.keys(objRow).length == 4, "objRow has length == 4: "+ Object.keys(objRow).length );
+			assert.ok( (objRow.PrincipalId && objRow.PrincipalType && objRow.Title && objRow.RoleNames), "All cols exist: "+ Object.keys(arrayResults[0]).toString() );
+			assert.ok( (typeof objRow.PrincipalType === 'number'), "typeof objRow.PrincipalType === 'number': "+ typeof objRow.PrincipalType );
+			assert.ok( (Array.isArray(objRow.RoleNames)), "Array.isArray(objRow.RoleNames): "+ Array.isArray(objRow.RoleNames) );
+			assert.ok( (objRow.RoleNames[0].Name), "objRow.RoleNames[0].Name exists: "+ objRow.RoleNames[0].Name );
+			assert.ok( getAsciiTableStr(arrayResults), `RESULTS:\n${getAsciiTableStr(arrayResults)}` );
+			done();
+		})
+		.catch(function(err){
+			assert.ok( (false), err );
+			done();
+		});
+	});
 }
 
 // ================================================================================================
@@ -1098,11 +1156,12 @@ QUnit.module( "REST Methods" );
 		var done = assert.async();
 		// TEST:
 		sprLib.rest({
-			url: "_api/lists/getbytitle('Employees')/items?$select=Id,Name,Manager/Title&$orderby=ID%20desc&$top=5&$expand=Manager"
+			url: "_api/lists/getbytitle('Employees')/items?$select=Name,Manager/Title&$orderby=ID%20desc&$top=5&$expand=Manager"
 		})
 		.then(function(arrayResults){
-			// NOTE: Running your own select results in raw results - sprLib only parses `queryCols` (hence 5 col shere and "unparsed" Manager/Title)
-			assert.ok( Object.keys(arrayResults[0]).length == 5, "arrayResults[0] has length == 5: "+ Object.keys(arrayResults[0]).length );
+			// NOTE: Running your own select results in raw results (sprLib only parses `queryCols`). Ex: `{"Title":"Brent Ely"}`
+			assert.ok( Object.keys(arrayResults[0]).length == 2, "arrayResults[0] has length == 2: "+ Object.keys(arrayResults[0]).length );
+			assert.ok( arrayResults[0].Manager, "arrayResults[0].Manager: "+ arrayResults[0].Manager );
 			assert.ok( arrayResults.length == 5, "arrayResults has length == 5: "+ arrayResults.length );
 			assert.ok( getAsciiTableStr(arrayResults), `RESULTS:\n${getAsciiTableStr(arrayResults)}` );
 			done();
@@ -1124,6 +1183,33 @@ QUnit.module( "REST Methods" );
 		.then(function(arrayResults){
 			assert.ok( Object.keys(arrayResults[0]).length == 3, "arrayResults[0] has length == 3: "+ Object.keys(arrayResults[0]).length );
 			assert.ok( (arrayResults[0].Manager.Title), "arrayResults[0].Manager.Title exists: "+ arrayResults[0].Manager.Title );
+			assert.ok( getAsciiTableStr(arrayResults), `RESULTS:\n${getAsciiTableStr(arrayResults)}` );
+			done();
+		})
+		.catch(function(err){
+			assert.ok( (false), err );
+			done();
+		});
+	});
+
+	// TEST: Complete, full URL
+	QUnit.test("sprLib.rest() ex: 'https://full.url.com/_api/web/sitegroups'", function(assert){
+		var done = assert.async();
+		// TEST:
+		sprLib.rest({
+			url: BASEURL+'/_api/web/sitegroups',
+			queryCols: {
+				title: { dataName:'Title' },
+				loginName: { dataName:'LoginName' },
+				editAllowed: { dataName:'AllowMembersEditMembership' }
+			}
+			,queryFilter:   "AllowMembersEditMembership eq true"
+			,queryOrderby:  "Title"
+			,queryLimit: 10
+		})
+		.then(function(arrayResults){
+			assert.ok( arrayResults.length > 0, "arrayResults is an Array and length > 0: "+ arrayResults.length );
+			assert.ok( (arrayResults[0].editAllowed), "arrayResults[0].editAllowed exists: "+ arrayResults[0].editAllowed );
 			assert.ok( getAsciiTableStr(arrayResults), `RESULTS:\n${getAsciiTableStr(arrayResults)}` );
 			done();
 		})
@@ -1195,6 +1281,109 @@ QUnit.module( "REST Methods" );
 		});
 	});
 	*/
+}
+
+// ================================================================================================
+QUnit.module( "REST > Parsing/Options Tests" );
+// ================================================================================================
+{
+	// NOTE: Parameterized QUnit Tests (!)
+	var arrObjTests = [
+		{ testDesc:"url:relative", urlPath:  "_api/lists/getbytitle('Site Assets')/items" },
+		{ testDesc:"url:absolute", urlPath: "/_api/lists/getbytitle('Site Assets')/items" },
+		{ testDesc:"url:relative", urlPath:  "_api/lists/getbytitle('Site Assets')/items?$select=ID" },
+		{ testDesc:"url:absolute", urlPath: "/_api/lists/getbytitle('Site Assets')/items?$select=ID" },
+		{ testDesc:"url:RESTROOT", urlPath: RESTROOT+"/_api/lists/getbytitle('Site Assets')/items?$select=ID" },
+		{ testDesc:"url:BASEURL+RESTROOT", urlPath: BASEURL+"/_api/lists/getbytitle('Site Assets')/items?$select=ID" },
+		{
+			testDesc: "query: queryCols",
+			urlPath:  "_api/lists/getbytitle('Site Assets')/items",
+			qryCols:  "ID",
+			arrTests: [
+				function(arrResults){ return arrResults.length > 0 },
+				function(arrResults){ return Object.keys(arrResults[0]).length == 1 },
+				function(arrResults){ return arrResults[0].ID }
+			]
+		},
+		{
+			testDesc: "query: queryCols + queryFilter",
+			urlPath: "_api/lists/getbytitle('Site Assets')/items",
+			qryCols: ['ID'],
+			qryFilter: "ID eq 10",
+			arrTests: [
+				function(arrResults){ return arrResults.length > 0 },
+				function(arrResults){ return Object.keys(arrResults[0]).length == 1 },
+				function(arrResults){ return arrResults[0].ID == 10 }
+			]
+		},
+		{
+			testDesc: "query: queryFilter + qryLimit",
+			urlPath: "_api/lists/getbytitle('Site Assets')/items",
+			qryFilter: "ID gt 10",
+			qryLimit: "5",
+			arrTests: [
+				function(arrResults){ return arrResults.length == 5 },
+				function(arrResults){ return arrResults[0].ID > 10 }
+			]
+		},
+		{
+			testDesc: "query: queryFilter",
+			urlPath: "_api/lists/getbytitle('Site Assets')/items",
+			qryFilter: "ID eq 10",
+			arrTests: [
+				function(arrResults){ return arrResults.length > 0 },
+				function(arrResults){ return arrResults[0].ID == 10 }
+			]
+		},
+		{
+			testDesc: "query: mixed $select and queryFilter",
+			urlPath: "_api/lists/getbytitle('Site Assets')/items?$select=ID",
+			qryFilter: "ID eq 10",
+			arrTests: [
+				function(arrResults){ return arrResults.length > 0 },
+				function(arrResults){ return arrResults[0].ID == 10 }
+			]
+		}
+
+		/*
+		{
+			testDesc : ""
+			urlPath  : RESTROOT+"/_api/lists/getbytitle('Site Assets')/items",
+			urlSelect: "",
+			qryCols  : null,
+			qryFilter: "",
+			qryLimit : ""
+		}
+		*/
+	];
+
+	QUnit.test("sprLib.rest() -> Battery of Parsing Tests (Total: "+arrObjTests.length+")", function(assert){
+		arrObjTests.forEach((objTest,idx)=>{
+			// A:
+			var done = assert.async();
+
+			// B: Set query object
+			var objRest = {};
+			if ( objTest.urlPath   ) objRest.url =  objTest.urlPath;
+			if ( objTest.urlSelect ) objRest.url += objTest.urlSelect;
+			if ( objTest.qryCols   ) objRest.queryCols   = objTest.qryCols;
+			if ( objTest.qryFilter ) objRest.queryFilter = objTest.qryFilter;
+			if ( objTest.qryLimit  ) objRest.queryLimit  = objTest.qryLimit;
+			if ( !objTest.arrTests ) objTest.arrTests = [function(arrResults){ return arrResults.length > 0 }];
+
+			// C: Execute test
+			sprLib.rest(objRest)
+			.then(function(arrResults){
+				objTest.arrTests.forEach((funcTest,idy) => assert.ok( funcTest(arrResults), (objTest.testDesc || "TEST "+idx)+" #"+idy+":\n"+funcTest.toString() ));
+				assert.ok(true, `RESULTS:\n${getAsciiTableStr(arrResults)}\n************************************************************\n`);
+				done();
+			})
+			.catch(function(err){
+				assert.ok( (false), err );
+				done();
+			});
+		});
+	});
 }
 
 // ================================================================================================
@@ -1273,108 +1462,11 @@ sprLib.rest({ restUrl:'../_api/web/GetByTitle' });
 */
 
 
-/*
-QUnit.test("sprLib.getItems() - onDone", function(assert){
-	var done = assert.async();
-	sprLib.getItems({
-		listName: 'Employees',
-		listCols: { id:{dataName:'ID'} },
-		queryLimit: 1,
-		onDone: function(){ assert.ok( true, "onDone fired!" ); done(); }
-	});
-});
-
-QUnit.test("sprLib.getItems() - with listCols", function(assert){
-	var done = assert.async();
-	sprLib.getItems({
-		listName: 'Employees',
-		listCols: {
-			id:       { dataName:'ID' },
-			name:     { dataName:'Name' },
-			badgeNum: { dataName:'Badge_x0020_Number' },
-			hireDate: { dataName:'Hire_x0020_Date', dispName:'Hire Date', dataFormat:'INTL' },
-			salary:   { dataName:'Salary' },
-			extn:     { dataName:'Extension' },
-			utilPct:  { dataName:'Utilization_x0020_Pct', dispName:'Util %' },
-			comments: { dataName:'Comments' }
-		},
-		queryLimit: 10,
-		onDone: function(arrayResults){
-			assert.ok( $.isArray(arrayResults), "onDone result is an array" );
-			assert.ok( arrayResults.length > 0, "arrayResults.length > 0" );
-			assert.ok( (arrayResults[0].__metadata && typeof arrayResults[0].__metadata !== 'undefined'), "arrayResults[0] is valid -> __metadata: "+ arrayResults[0].__metadata );
-			assert.ok( (arrayResults[0].id         && typeof arrayResults[0].name       !== 'undefined'), "arrayResults[0] is valid -> Id: "+arrayResults[0].id+" / Title: "+arrayResults[0].name );
-			// TODO: Move to the (as yet undone) MODEL TEST section
-
-			/ *
-			QUnit.test("sprLib.model('Emp').data() method", function(assert){
-				assert.ok( $.isArray(sprLib.model('Employees').data()), "sprLib.model().add().data() is an array" );
-				assert.ok( $.isArray(sprLib.model('Employees').data('array')), "sprLib.model().add().data('array') is an array" );
-				assert.ok( (typeof sprLib.model('Employees').data('object') === 'object'), "sprLib.model().add().data('object') is an object" );
-			});
-			* /
-			done();
-		}
-	});
-});
-
-QUnit.test("sprLib.getItems() - w/o listCols", function(assert){
-	var done = assert.async();
-	sprLib.getItems({
-		listName: 'Employees',
-		queryLimit: 10,
-		onDone: function(arrayResults){
-			assert.ok( $.isArray(arrayResults), "onDone result is an array" );
-			assert.ok( arrayResults.length > 0, "arrayResults.length > 0" );
-			assert.ok( ( arrayResults[0].Id ), "arrayResults[0] is valid - Id: "+arrayResults[0].Id );
-			QUnit.test("sprLib.model('Emp').data() method", function(assert){
-				assert.ok( $.isArray(sprLib.model('Employees').data()), "sprLib.model().add().data() is an array" );
-				assert.ok( $.isArray(sprLib.model('Employees').data('array')), "sprLib.model().add().data('array') is an array" );
-				assert.ok( (typeof sprLib.model('Employees').data('object') === 'object'), "sprLib.model().add().data('object') is an object" );
-			});
-			done();
-		}
-	});
-});
-
-QUnit.test("sprLib.getItems() - dataFunc listCols", function(assert){
-	var done = assert.async();
-	sprLib.getItems({
-		listName: 'Employees',
-		listCols: {
-			name:     { dataName:'Name' },
-			badgeNum: { dataName:'Badge_x0020_Number' },
-			funcTest: { dataFunc:function(objItem){ return objItem.Name +' ('+ objItem.Badge_x0020_Number+')' } }
-		},
-		queryLimit: 10,
-		onDone: function(arrayResults){
-			assert.ok( $.isArray(arrayResults), "onDone result is an array" );
-			assert.ok( arrayResults.length > 0, "arrayResults.length > 0" );
-			assert.ok( ( arrayResults[0].badgeNum ), "arrayResults[0] is valid - badgeNum: " + arrayResults[0].badgeNum );
-			assert.ok( ( arrayResults[0].funcTest ), "arrayResults[0] is valid - funcTest: " + arrayResults[0].funcTest );
-			done();
-		}
-	});
-});
-*/
-
 // ================================================================================================
 //QUnit.module( "Binding / Forms" );
 // ================================================================================================
 // sprLib.model('Employees').syncItem()
 
-// WORKS
-/*
-sprLib.model('junk').add({
-	listName: '/sites/dev/_api/web/sitegroups',
-	listCols: {
-		title: { dataName:'Title' },
-		loginName: { dataName:'LoginName' },
-		editAllowed: { dataName:'AllowMembersEditMembership' }
-	},
-	onDone: function(data){ console.table(data) }
-});
-*/
 
 /*
 some API calls require an arguamnet (group/pr0file rest endpoints ets), require an auth toekn an POST!